:shield: separated sendgrid ui logic to backend for security
