+++ marketplace/backend/api/v1/services/sendGridService.js
@@ -1,0 +1,25 @@
+import sgMail from '@sendgrid/mail';
+import dotenv from 'dotenv';
+
+dotenv.config({ path: '../../../.env' });
+
+sgMail.setApiKey(process.env.SENDGRID_API_KEY);
+
+async function sendEmail(to, subject, htmlContent) {
+  const msg = {
+    to: to,
+    from: { email: "srinjoy_chakravarty@blockapps.net", name: "Blockapps.net" },
+    subject: subject,
+    html: htmlContent
+  };
+
+  try {
+    await sgMail.send(msg);
+    console.log('Email sent successfully!');
+  } catch (error) {
+    console.error('Error sending email:', error);
+    throw error;
+  }
+}
+
+export { sendEmail };
