+++ marketplace/backend/package.json
@@ -50,6 +50,7 @@
   "private": true,
   "dependencies": {
     "@hapi/joi": "^17.1.1",
+    "@sendgrid/mail": "^7.7.0",
     "aws-sdk": "^2.1284.0",
     "babel-plugin-module-resolver": "^4.1.0",
     "blockapps-rest": "8.1.0",
